+++ src/Vector/BLF/LinShortOrSlowResponse2.h
@@ -28,7 +28,6 @@
 #include <Vector/BLF/AbstractFile.h>
 #include <Vector/BLF/LinDatabyteTimestampEvent.h>
 #include <Vector/BLF/ObjectHeader.h>
-#include <Vector/BLF/VectorTypes.h>
 
 #include <Vector/BLF/vector_blf_export.h>
 
@@ -43,14 +42,14 @@ struct VECTOR_BLF_EXPORT LinShortOrSlowResponse2 final : ObjectHeader, LinDataby
 
     void read(AbstractFile & is) override;
     void write(AbstractFile & os) override;
-    DWORD calculateObjectSize() const override;
+    uint32_t calculateObjectSize() const override;
 
     /**
      * @brief number of valid response bytes
      *
      * The number of response bytes.
      */
-    ULONG numberOfRespBytes {};
+    uint32_t numberOfRespBytes {};
 
     /**
      * @brief the response bytes (can include the checksum)
@@ -58,7 +57,7 @@ struct VECTOR_BLF_EXPORT LinShortOrSlowResponse2 final : ObjectHeader, LinDataby
      * The response bytes (can include the
      * checksum).
      */
-    std::array<BYTE, 9> respBytes {};
+    std::array<uint8_t, 9> respBytes {};
 
     /**
      * @brief non-zero, if the response was too slow
@@ -66,7 +65,7 @@ struct VECTOR_BLF_EXPORT LinShortOrSlowResponse2 final : ObjectHeader, LinDataby
      * Non-zero, if the response was too
      * slow; otherwise zero.
      */
-    BYTE slowResponse {};
+    uint8_t slowResponse {};
 
     /**
      * @brief non-zero, if the response was interrupted by a sync break
@@ -75,10 +74,10 @@ struct VECTOR_BLF_EXPORT LinShortOrSlowResponse2 final : ObjectHeader, LinDataby
      * interrupted by a sync break;
      * otherwise zero.
      */
-    BYTE interruptedByBreak {};
+    uint8_t interruptedByBreak {};
 
     /** reserved */
-    BYTE reservedLinShortOrSlowResponse1 {};
+    uint8_t reservedLinShortOrSlowResponse1 {};
 
     /**
      * @brief Exact baudrate of the header in bit/sec
@@ -86,7 +85,7 @@ struct VECTOR_BLF_EXPORT LinShortOrSlowResponse2 final : ObjectHeader, LinDataby
      * Event's baudrate measured in
      * header [in bits/sec]
      */
-    DOUBLE exactHeaderBaudrate {};
+    double exactHeaderBaudrate {};
 
     /**
      * @brief Early stop bit offset for UART timestamps in ns
@@ -95,10 +94,10 @@ struct VECTOR_BLF_EXPORT LinShortOrSlowResponse2 final : ObjectHeader, LinDataby
      * header for UART timestamps
      * [in ns]
      */
-    DWORD earlyStopbitOffset {};
+    uint32_t earlyStopbitOffset {};
 
     /** reserved */
-    DWORD reservedLinShortOrSlowResponse2 {};
+    uint32_t reservedLinShortOrSlowResponse2 {};
 };
 
 }
