+++ src/Vector/BLF/LinSpikeEvent.h
@@ -25,7 +25,6 @@
 
 #include <Vector/BLF/AbstractFile.h>
 #include <Vector/BLF/ObjectHeader.h>
-#include <Vector/BLF/VectorTypes.h>
 
 #include <Vector/BLF/vector_blf_export.h>
 
@@ -45,24 +44,24 @@ struct VECTOR_BLF_EXPORT LinSpikeEvent final : ObjectHeader {
 
     void read(AbstractFile & is) override;
     void write(AbstractFile & os) override;
-    DWORD calculateObjectSize() const override;
+    uint32_t calculateObjectSize() const override;
 
     /**
      * @brief application channel
      *
      * Channel number where the frame sent/received.
      */
-    WORD channel {};
+    uint16_t channel {};
 
     /** reserved */
-    WORD reservedLinSpikeEvent {};
+    uint16_t reservedLinSpikeEvent {};
 
     /**
      * @brief the spike's width
      *
      * Spike length in microseconds
      */
-    ULONG width {};
+    uint32_t width {};
 };
 
 }
